src/main/java/com/credibanco/bancinc/BancincApplication.java
package com.credibanco.bancinc;

import org.springframework.boot.SpringApplication;
import org.springframework.boot.autoconfigure.SpringBootApplication;
import org.springframework.boot.builder.SpringApplicationBuilder;
import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;

@SpringBootApplication
public class BancincApplication extends SpringBootServletInitializer  {

	public static void main(String[] args) {
		SpringApplication.run(BancincApplication.class, args);
	}

	@Override
	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
		return application.sources(BancincApplication.class);
	}

}
